# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Codex2API
  description: |
    OpenAI-compatible API proxy for Codex with enhanced features including:
    - Multi-account pooling with automatic rotation
    - Intelligent request routing and load balancing
    - Real-time usage tracking and monitoring
    - Administrative management interface
    - Request/response translation between OpenAI and Codex formats
    - Streaming and non-streaming chat completions
    - Function calling support
    - Tool calling support
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/codex2api
    email: support@codex2api.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/v1
    description: Local development server (recommended versioned base URL)
  - url: http://localhost:8080
    description: Local development server (root-path compatibility base URL)
  - url: https://api.codex2api.local/v1
    description: Production server (recommended versioned base URL)
  - url: https://api.codex2api.local
    description: Production server (root-path compatibility base URL)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Enter your API key with Bearer prefix.
        Example: "Bearer sk-..."
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - type
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - invalid_request
                - invalid_api_key
                - missing_api_key
                - insufficient_scope
                - invalid_parameter
                - missing_field
                - invalid_field_type
                - invalid_field_format
                - context_length_exceeded
                - unsupported_model
                - rate_limit_reached
                - server_error
                - service_unavailable
                - upstream_error
                - upstream_timeout
                - resource_not_found
                - resource_conflict
            message:
              type: string
              description: Human-readable error message
            type:
              type: string
              description: Error category
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found_error
                - rate_limit_error
                - server_error
                - upstream_error
            details:
              description: Additional error context
              oneOf:
                - type: object
                  additionalProperties: true
                - type: array
                  items:
                    type: object
                    additionalProperties: true
...
    ResponsesRequest:
      type: object
      required:
        - model
        - input
      properties:
        model:
          type: string
          description: ID of the model to use
        input:
          type: array
          description: Input items for the response
          items:
            type: object
        instructions:
          type: string
        max_output_tokens:
          type: integer
          minimum: 1
          maximum: 65536
        temperature:
          type: number
          minimum: 0
          maximum: 2
        top_p:
          type: number
          minimum: 0
          maximum: 1
        stream:
          type: boolean
        stop:
          type: string
        user:
          type: string
        reasoning:
          type: object
          properties:
            effort:
              type: string
              enum: [low, medium, high, xhigh]
        service_tier:
          type: string
          enum: [auto, default, fast]
        store:
          type: boolean
        truncation:
          type: string
          enum: [auto, disabled]
        previous_response_id:
          type: string
        tools:
          type: array
        tool_choice:
          type: object
    ResponsesResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          enum: [response]
        created_at:
          type: integer
        status:
          type: string
          enum: [in_progress, completed, failed, cancelled, incomplete]
        error:
          $ref: '#/components/schemas/Error'
        model:
          type: string
        output:
          type: array
        usage:
          type: object
          properties:
            input_tokens:
              type: integer
            output_tokens:
              type: integer
            total_tokens:
              type: integer
        incomplete_details:
          type: object
    ImagesGenerationRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          description: Text prompt used to generate images
        model:
          type: string
          description: Image model ID
          default: gpt-image-2
          example: gpt-image-2
        response_format:
          type: string
          description: Output payload format
          enum: [b64_json, url]
          default: b64_json
        size:
          type: string
          description: Requested output size
          example: 1024x1024
        quality:
          type: string
          description: Requested output quality
        background:
          type: string
          description: Background style
        output_format:
          type: string
          description: Image output format (e.g. png/jpeg/webp)
        output_compression:
          type: integer
          description: Compression level for supported output formats
        moderation:
          type: string
          description: Moderation mode
        n:
          type: integer
          description: Ignored for compatibility and not forwarded upstream
    ImagesEditsRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          description: Text prompt describing the desired edit result
        model:
          type: string
          description: Image model ID
          default: gpt-image-2
          example: gpt-image-2
        image:
          description: Base image input (single string URL/data URL or array form)
          oneOf:
            - type: string
            - type: array
              items:
                oneOf:
                  - type: string
                  - type: object
                    properties:
                      image_url:
                        type: string
        images:
          type: array
          description: Additional image inputs (JSON mode)
          items:
            oneOf:
              - type: string
              - type: object
                properties:
                  image_url:
                    type: string
                  url:
                    type: string
        mask:
          description: Optional mask image URL/data URL
          oneOf:
            - type: string
            - type: object
              properties:
                image_url:
                  type: string
                url:
                  type: string
        response_format:
          type: string
          description: Output payload format
          enum: [b64_json, url]
          default: b64_json
        size:
          type: string
          description: Requested output size
          example: 1024x1024
        quality:
          type: string
          description: Requested output quality
        background:
          type: string
          description: Background style
        output_format:
          type: string
          description: Image output format (e.g. png/jpeg/webp)
        output_compression:
          type: integer
          description: Compression level for supported output formats
        input_fidelity:
          type: string
          description: Input fidelity mode
        moderation:
          type: string
          description: Moderation mode
        n:
          type: integer
          description: Ignored for compatibility and not forwarded upstream
    ImageData:
      type: object
      properties:
        b64_json:
          type: string
          description: Base64-encoded image data (without data URL prefix)
        url:
          type: string
          description: Image URL (or data URL when URL mode is requested)
        revised_prompt:
          type: string
          description: Prompt rewritten by the model, if provided upstream
    ImagesResponse:
      type: object
      required:
        - created
        - data
      properties:
        created:
          type: integer
          description: Unix timestamp
        data:
          type: array
          items:
            $ref: '#/components/schemas/ImageData'
paths:
...
  /responses:
    post:
      summary: Create response
      description: |
        Creates a response using the native Codex Responses API.
        Provides direct access to Codex functionality.
      operationId: createResponse
      tags:
        - Responses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponsesRequest'
      responses:
        '200':
          description: Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsesResponse'
            text/event-stream:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /images/generations:
    post:
      summary: Generate images
      description: |
        OpenAI-compatible images generation endpoint.
        Default model is `gpt-image-2`.
      operationId: createImageGeneration
      tags:
        - Images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImagesGenerationRequest'
      responses:
        '200':
          description: Images response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImagesResponse'
        '400':
          description: Invalid request or unsupported model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /images/edits:
    post:
      summary: Edit images
      description: |
        OpenAI-compatible images edit endpoint.
        Current implementation supports JSON requests only.
        Multipart requests return 400 with a clear unsupported message.
      operationId: createImageEdit
      tags:
        - Images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImagesEditsRequest'
          multipart/form-data:
            schema:
              type: object
              description: Reserved for future support; currently returns 400.
      responses:
        '200':
          description: Images response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImagesResponse'
        '400':
          description: Invalid request or unsupported model/content type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...